Seaport Tracker - individual school project
Reads a .txt file of various seaports with their ships (cargo and passenger) docked or queued with crew information. The program 
reads and scans the “world” .txt file and can display the seaports, cargo and passenger ships, and all the crewmen of that world written in the .txt file. 


The program can sort the data alphabetically or by key and can be displayed in regular output text or through a tree. 
The level of inheritance are shown by the three ‘main’ classes using the titled SeaPortTrackerOne, SeaPortTrackerTwo, and SeaPortTrackerThree.